+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/domain/model/params/GetInstallmentParams.kt
@@ -1,8 +1,10 @@
 package com.mercadopago.sdk.android.coremethods.domain.model.params
 
+import java.math.BigDecimal
+
 internal data class GetInstallmentParams(
     val productId: String? = null,
     val bin: Int? = null,
     val processingMode: String? = null,
-    val amount: Long? = null,
+    val amount: BigDecimal? = null,
 )
